refactor: move model to a not circle place
