+++ autogpt_platform/backend/backend/executor/manager.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING, Any, Generator, Optional, TypeVar, cast
 from redis.lock import Lock as RedisLock
 
 from backend.blocks.basic import AgentOutputBlock
-from backend.data.execution import (
+from backend.data.model import (
     GraphExecutionStats,
     NodeExecutionStats,
     convert_graph_execution_stats,
